vectorstores: Weaviate add default id field
Via 'queryAttrs' you can specify aditional field to be returned with a
simularity search. In Weaviate you can retrieve the id via the
'_additional' field. With the current setup it is not possible to set
the '_additional' due to the manual mapping in 'createFields' function
to retrieve the score (certainty).